WIP!!! have to push otherwise data will be lost with windows reinstallation

+++ src/DayTime.cpp
@@ -80,7 +80,7 @@ DayTime::DayTime(float timeInHours)
 {
   long sgn = fsign(timeInHours);
   timeInHours = fabsf(timeInHours);
-  totalSeconds = sgn * static_cast<long>(roundf(timeInHours * 60.0f * 60.0f));
+  totalSeconds = sgn * round(timeInHours * 60 * 60);
 }
 
 int DayTime::getHours() const
